fixed issues with draw instruction, and more
Fixed bug regarding op_7XNN(VF flag shouldn't be set even if a carry
were to happen).
Fixed small VF bug in subtraction instructions (Flag shoul be set if
a == b in a-b).
The draw instruction bugs were 2: Firstly, i was left shifting by 8
instead of 7, and secondly, i forgot to right shift back. The pygame
logic checked if the value in the register was different from 0, and i
was only expecting 1 or 0 values. By not shifting (and XORing), i was
getting other values like 128 and 64.